requirements.txt
av==12.3.0
certifi==2024.7.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.3.1
faster-whisper==1.0.3
ffmpeg-python==0.2.0
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
future==1.0.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.1
packaging==24.1
protobuf==5.27.3
PyYAML==6.0.1
requests==2.32.3
sympy==1.13.1
tokenizers==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
